+++ tests/e2e/cases/generated_c_dce_const/in.ts
@@ -492,6 +492,9 @@ const unused_promise_resolve_object_symbols_find_call = Promise.resolve(Object.g
 const unused_promise_resolve_object_symbols_find_index_call = Promise.resolve(Object.getOwnPropertySymbols({ dead_promise_resolve_object_symbols_find_index: 1 }).findIndex(() => "dead_promise_resolve_object_symbols_find_index_callback".length > 0));
 const unused_promise_resolve_empty_array_reduce_call = Promise.resolve([].reduce((acc: string) => acc + "dead_promise_resolve_empty_array_reduce_callback", "dead_promise_resolve_empty_array_reduce_initial"));
 const unused_promise_resolve_object_symbols_reduce_right_call = Promise.resolve(Object.getOwnPropertySymbols({ dead_promise_resolve_object_symbols_reduce_right: 1 }).reduceRight((acc: string) => acc + "dead_promise_resolve_object_symbols_reduce_right_callback", "dead_promise_resolve_object_symbols_reduce_right_initial"));
+const unused_promise_resolve_object_symbols_slice_length = Promise.resolve(Object.getOwnPropertySymbols({ dead_promise_resolve_object_symbols_slice_length: 1 }).slice().length);
+const unused_promise_resolve_object_symbols_to_reversed_length = Promise.resolve(Object.getOwnPropertySymbols({ dead_promise_resolve_object_symbols_to_reversed_length: 1 }).toReversed().length);
+const unused_promise_resolve_object_symbols_keys_length = Promise.resolve(Object.getOwnPropertySymbols({ dead_promise_resolve_object_symbols_keys_length: 1 }).keys().length);
 const unused_promise_resolve_object_extensible_call = Promise.resolve(Object.isExtensible({ dead_promise_resolve_object_extensible: 1 }));
 const unused_promise_resolve_object_sealed_call = Promise.resolve(Object.isSealed({ dead_promise_resolve_object_sealed: 1 }));
 const unused_promise_resolve_object_frozen_call = Promise.resolve(Object.isFrozen({ dead_promise_resolve_object_frozen: 1 }));
@@ -3864,6 +3867,8 @@ function usedLocal(value: number): number {
     Object.getOwnPropertySymbols(Object.freeze({ local_dead_object_symbols_find_last: 1 })).findLast(() => "local_dead_object_symbols_find_last_callback".length > 0);
     Object.getOwnPropertySymbols(Object.freeze({ local_dead_object_symbols_find_last_index: 1 })).findLastIndex(() => "local_dead_object_symbols_find_last_index_callback".length > 0);
     Object.getOwnPropertySymbols(Object.freeze({ local_dead_object_symbols_reduce: 1 })).reduce((acc: string) => acc + "local_dead_object_symbols_reduce_callback", "local_dead_object_symbols_reduce_initial");
+    Object.getOwnPropertySymbols(Object.freeze({ local_dead_object_symbols_values_length: 1 })).values("local_dead_object_symbols_values_ignored").length;
+    Object.getOwnPropertySymbols(Object.freeze({ local_dead_object_symbols_entries_length: 1 })).entries("local_dead_object_symbols_entries_ignored").length;
     Object.getOwnPropertyDescriptor("local_dead_object_primitive_descriptor", "0");
     Object.getOwnPropertyDescriptors("local_dead_object_primitive_descriptors");
     Object.hasOwn("local_dead_object_primitive_has_own", "0");